Add -I$(MPAS_LIBDIR) to INCLDIR when building the MPAS-A dycore in CAM
This commit adds the path of the MPAS-A library, which is also the path to
MPAS-A *.mod files, to INCLDIR when building CAM with the MPAS-A dycore.
This commit also adds a dependency of dyn_comp.o on $(MPAS_LIBDIR)/libmpas.a
to ensure that the MPAS-A dycore has been fully build before we attempt to
use any MPAS-A .mod files in dyn_comp.