Declare MPAS-A dycore library as a PHONY target
The MPAS-A dycore library target in scripts/Tools/Makefile was previously
named libmpas.a (i.e., the name of the actual library file). Because this
target was not declared as PHONY, as long as libmpas.a existed, modified
MPAS-A dycore source files were not re-compiled because the recipe for
libmpas.a was not executed.

To ensure that any modified MPAS-A source files are recompiled and archived
into the libmpas.a library, the name of the dycore library target has been
changed from 'libmpas.a' to 'libmpas', and libmpas has furthermore been
declared as PHONY.